Added component diagram and moved class diagram for app
Class diagram is now under design spec as it is a design issue not a
project plan.

Add component diagram for the android app. Ben if you could take a look
at it tell me if it needs improvements, I went by what I could see on
the church heating doc.